refactor(frontend): move document domain types into documents module
