+++ frontend/src/modules/documents/domain/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./document.types";
