+++ frontend/src/modules/documents/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./domain";
